Refactor Spark cluster creation and make cluster creation optional for ADAM.
Factors the Spark cluster creation out of the ADAM pipeline into it's own
package (toil_scripts.spark_utils). Restructures the ADAM pipeline to declare
the pipeline statically, and renames it with a more descriptive name. The
static pipeline definition for the ADAM preprocessing pipeline now allows
a user to pass a master IP for a static Spark cluster. Resolves #137.
