+++ src/toil_scripts/__init__.py
@@ -1,3 +1,7 @@
+import logging
+
+_log = logging.getLogger(__name__)
+
 def download_from_s3_url(file_path, url):
     from urlparse import urlparse
     from boto.s3.connection import S3Connection
@@ -9,6 +13,7 @@ def download_from_s3_url(file_path, url):
                                "s3://BUCKET/KEY. '%s' is not." % url)
         bucket = s3.get_bucket(parsed_url.netloc)
         key = bucket.get_key(parsed_url.path[1:])
+        _log.info("From URL %s, parsed --> %s, bucket %s, key %s", url, parsed_url, bucket, key)
         key.get_contents_to_filename(file_path)
     finally:
         s3.close()
